Update patch set 1
Patch Set 1:

> There are some whitespace issues (trailing spaces) in
 > TEST_PLAN_INFO - how strict do we want to be on things like this,
 > or do we just accept and fix it?

There are similar issues with most of the dir.info.txt files, and there is a superfluous "it" in the .odt. Last line should be:


 > There are some whitespace issues (trailing spaces) in
 > TEST_PLAN_INFO - how strict do we want to be on things like this,
 > or do we just accept and fix it?

 > There are some whitespace issues (trailing spaces) in
 > TEST_PLAN_INFO - how strict do we want to be on things like this,
 > or do we just accept and fix it?

There are similar issues with the dir.info.txt files (all but section 05). And the last line of the .odt should be:

3. To pass a test case, all individual metrics would be expected to 

Patch-set: 1