Proposing a Dovetail test suite structure for IPv6
Adding the first draft of the IPv6 test suite.
This patch attempts to start the documentation according to the
proposal on https://wiki.opnfv.org/pages/viewpage.action?pageId=6827269
The text is neither complete nor accurate at this stage, this patch
is intended to establish a concrete dialog for how we structure our
test suites and reference test cases.

Change-Id: I156ddafa3e8b7630390523864490426fc742eb3a
Signed-off-by: ChristopherPrice <christopher.price@ericsson.com>