IPv6 test cases.
We are keeping the test methodology simple and easy.
We are reusing existing tests as much as we can from
IPv6 projects and Tempest API and Scenario tests, which
are run in FuncTest and Yardstick and integrated into
OPNFV's CI/CD environment.

Please note that previous templates are not applicable
any more because those templates and contents do not
match IPv6 test cases.

L3VPN test cases will be submitted in a separate patch.

Change-Id: I3e6c4962567ea0b79875316c9af535ae17516533
Signed-off-by: Bin Hu <bh526r@att.com>