IPv6 test cases, document example by the way
We are keeping the test methodology simple and easy.
We are reusing existing tests as much as we can from
IPv6 projects and Tempest API and Scenario tests, which
are run in FuncTest and Yardstick.

hongbo:
example of test case specification document are
ipv6.tc001.specification.rst
ipv6.tc026.specification.rst
considerations:
1,suitable with our wiki test case document https://wiki.opnfv.org/pages/viewpage.action?pageId=6827269
2,suitable with dovetail test tool https://wiki.opnfv.org/pages/viewpage.action?pageId=7766287

comments are welcomed to make our document example

Change-Id: I7e47c73c0f02a407df278c93a3da2142b5cc86e5
Signed-off-by: hongbo tian <hongbo.tianhongbo@huawei.com>