IPv6 test cases.
We are keeping the test methodology simple and easy.
We are reusing existing tests as much as we can from
IPv6 projects and Tempest API and Scenario tests, which
are run in FuncTest and Yardstick and integrated into
OPNFV's CI/CD environment.

Please note that previous templates are not applicable
any more because those templates and contents do not
match IPv6 test cases.

L3VPN test cases will be submitted in a separate patch.

Change-Id: I7e47c73c0f02a407df278c93a3da2142b5cc86e5
Signed-off-by: Bin Hu <bh526r@att.com>