Update patch set 3
Patch Set 3:

(8 comments)

Some comments on the document inline.
As the test specification is usually quite detailed and would contain  a lot of text I suggest moving away from a table format.  The spec itself should be able to be read and distributed as a stand alone document.

Patch-set: 3
Label: Code-Review=0