Update patch set 1
Patch Set 1:

(4 comments)

Looking good Leo.  
A few comments inline from a fast glace and some areas we should work on.  (Also the files need commenting)

Can we discuss the directory structure on the upcoming Dovetail call to make sure we are all on the same page regarding framework and test case structures in the Dovetail repo?

Patch-set: 1
Label: Code-Review=0