Update patch set 2
Patch Set 2: Code-Review-1

The request we have from the C&C committee is that all test cases need to be described in enough detail to be run manually.  I cannot see how to execute these test cases manually based on this test plan and specification.
Suggest we move away from the table format and begin to document the test plan, design and specification according to the request from the C&C committee.

Patch-set: 2