Update patch set 3
Patch Set 3: Code-Review-1

I have some problem with the structure we have here.
The file "total.yml" is in the cert directory even though as I understand it, it is not part of the cert activity but more a list of "potential" test cases.
Maybe either move total.yml out of cert or rename it to "proposed_tests" or some such thing to clarify the purpose.  It is certainly not the total certification suite.

Patch-set: 3