Update patch set 1
Patch Set 1:

Thank you all for the explanation, make sense if we create a new container for each testcase so that we would have a clear env to run the new case. just feel this will be time-consuming if we run prepare_env for each container(or each test case);  if we end up with having a lot of testcases, it would certainly be a time-consuming part that we might want to re-visit in the future.
I'll abandon this patch and mark the jira issue as resolved for now.

Patch-set: 1
Reviewer: Gerrit User 2667 <2667@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0