[RFC]Add dockerfile to support dovetail in CentOS
This diff introduces an optional dovetail dockerfile based on CentOS.
It provides users to have an alternative way to execute dovetail.
It should be just a optional, so we should keep docker/Dockerfile
as official one with dockerhub's image and we do not build docker images
in CI phase for now.

Change-Id: Ia694bcd49f65ff2bc175c1da3b42fbfa19b022a1
Signed-off-by: Tomofumi Hayashi <tohayash@redhat.com>