Update patch set 7
Patch Set 7: Code-Review+2

I am going to merge this as a draft in order that others can pick up JIRA tickets and begin to work against the document in parallel.
Going to create another patch to continue to structure the rest of the document and collect comments.  Will refer to this patch as a reference.

Patch-set: 7
Label: Code-Review=+2