dovetail tool: check push_to_db success or fail
JIRA: DOVETAIL-380

If Functest fails to push results to db, there is just an error log in functest.log
Dovetail needs to know the fail.
Dovetail will get the results from db with build_tag and testcase.
If it can get the results data, it means push_to_db success.

Change-Id: If03bdf0ceb2f1a9d422381cdf06e1fd207358bd5
Signed-off-by: xudan <xudan16@huawei.com>