Update patch set 1
Patch Set 1:

How do we proceed in this work? Trevor's suggestion makes sense to me. I would only add that, in Trevor's separation of 3 documents, the first would be for test case developers and not the users (i.e. testers). I will volunteer to draft the Testing Guide, in a separate patch for now, and we can decide on how to merge later.

Patch-set: 1