Update patch set 1
Patch Set 1:

(4 comments)

Updated with comments.
WRT requiring more documentation.  I completely agree that we need developer facing documentation.  This document is being written explicitly user facing and not intended as a design document.  I would not envision a CVP user shoudl be expected to read the design document to gather an understanding of how the CVP has been built or why it behaves the way it does.  Thus replication will be needed in some way for internal an external audiences.

Patch-set: 1
Label: Verified=0