Update patch set 1
Patch Set 1:

(7 comments)

I made some suggestions along the whole document. My main concern is the mixed/complex structure (see Line 201 comment and suggestion). If we can agree on putting the section on preparing the system and the section on conducting tests back-to-back together, then I'll be happy to submit a patch for that portion.

Patch-set: 1
Label: Verified=0