Committing the test case requirements specification to GIT for committer review.
This doucment needs to be +2'd by the majority of committers prior to merging
to ensure we are all in agreement with regard to the test case requirements.
This is a pre-requisite to test case selection for the CVP.

Updated by dneary 2017-04-25: List formatting, and reorganising requirements for installer and stack composition.

Change-Id: Iabaf11bc9479a20a58f985233d460013e4690d7c
Signed-off-by: ChristopherPrice <christopher.price@ericsson.com>