Update patch set 5
Patch Set 5: Code-Review+2

Double checked with Trevor and based on the etherpad I will push this as a structure of docs according to https://etherpad.opnfv.org/p/cvp_tooling_sketch.
The docs ARE NOT COMPLETE, however getting the structure in place allows us to spread the work out.

Patch-set: 5
Label: Code-Review=+2