Update patch set 2
Patch Set 2: Code-Review+2

I really like the test description here, I think all tests should have something similar. This is a good addition to the test suite, but I have a question (as detailed elsewhere) about how we mark tests as required vs optional, and what the consequences of optional test failures is for products. I am happy to approve this as-is as an example of a test case description, but continue to have concerns about including this in the CVP required tests.

Patch-set: 2
Label: Code-Review=+2