Update patch set 1
Patch Set 1:

Will review in detail, some basics.
Best practice for images is that all images are stored in a ./dovetail/docs/images/ directory.
There are a number of whitespace error in the logs.
I don't know that we should list for instance all the linux versions, tool versions are important though.
For the instruction I don't think we should be referring to version tags, we should be explicit in the syntax of the commands needed as an instruction.

Patch-set: 1
Label: Code-Review=0