Proposing to move "post conditions" as an optional topic for a test case.
It is not clear that there are post conditions to be documented for all test cases.
In the case a test cse creates an image and deletes it, there is littöe need for
a post condition to be documented.  If a test case establishes a network to be
used for a subsequent test ina  sequence this information should be clearly dcumented.

Change-Id: Ifb4bb45f240921c180077ab7273988d973417f6f
Signed-off-by: ChristopherPrice <christopher.price@ericsson.com>