Update patch set 1
Patch Set 1: Code-Review-1

I agree with George, this needs a lot more detailed description for the test cases (pre-requisites, set-up, tear-down, pass/fail criteria). Also, I notice that the file contains "Mandatory CVP Test Areas" as a heading - does this mean that you are proposing, in this patch, that these tests are to be included as mandatory tests? I think adding the test cases, and proposing that they be mandatory, should be done in 2 steps. I also believe that we are tracking mandatory/optional tests in a different place, as you can see in Gerrit #34343

Patch-set: 1
Label: Code-Review=-1