Update patch set 6
Patch Set 6:

I see all the alias are not included in testscope, so the question is if to include or not in document, in CVP tool, we have to include the alias testcases, since it's refstack requires and how functest works. if not included in doc and only included in CVP tool codes, at least it's OK for me, if both included in doc and codes, it's also OK. and suggest to keep the same in two documents, i.e, testscope and testarea doc(e.g. https://git.opnfv.org/dovetail/tree/docs/testing/user/testspecification/vimoperationsnetwork/index.rst)

Patch-set: 6
Reviewer: Gerrit User 371 <371@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0