I would like to nominate Georg Kunz to become a committer in the Dovetail project. Georg's nomination is motivated by his commitment to the project and his high quality work in helping the Dovetail sdnvpn test suite development, test doc patch review and meeting participations.
I would like to nominate Zenghui to become a committer in the Dovetail
project. Zenghui's nomination is motivated by his commitment to the
project and his contributions in dovetail tool and multi-SUT testing.

I would like to nominate Wanglei to become a committer in the Dovetail
project. Wanglei's nomination is motivated by his commitment to the project
and his consistent long time contribution to the Dovetail test framework
development, test suite patch review and meeting participations.

I would like to nominate Lincoln to become a committer in the Dovetail
project. Lincoln's nomination is motivated by his commitment to the project
and his expertise in multi-vendor testing, compliance testing and test doc
patch review and meeting participations.

I would like to nominate Trevor Cooper to become a committer in the Dovetail
project. Trevor's nomination is motivated by his commitment to the project
and his experiences in the testing working group, performance testing and
consistent contributions in reviews and meeting participations.

I would like to nominate Fuqiao to become a committer in the Dovetail project.
Fuqiao's nomination is motivated by her commitment to the project and her
contributions in developing the high availability test area, her experiences
in bringing service provider requirements and perspective, and her
contributions in reviews and meeting participations.

Change-Id: Ia8a4ed3af7219c5376e0dd281ed11a51059c6fd4
Signed-off-by: Wenjing Chu <chu.wenjing@gmail.com>