Update patch set 4
Patch Set 4: Code-Review-1

(3 comments)

Good detail in the test descriptions requiring a little improvement.  I did not re-comment other comments.
I think we should identify which tests are mandatory and which are optional.  HA in general is not optional, so I don't think these should all be optional tests.
Also the tests need to reflect the "danube" release, not describing any changes we might want to make to our release criteria in the future.

Patch-set: 4
Label: Code-Review=-1