Update patch set 4
Patch Set 4: Code-Review-1

(7 comments)

A lot of work is needed to complete this document.  The second half is still mostly empty.

In general the instructions need to be more concise, we need to identify the procedures clearly (rather than stating we don't know yet) and have a labeling scheme in place.

Patch-set: 4
Label: Code-Review=-1