Update patch set 5
Patch Set 5: Code-Review-1

(23 comments)

A few comments.  There are still a number of items in this that require some discussion/decisions to be taken.  Also the build is failing due to white-spaces that should be removed.
Suggest to remove the developer facing documentation to a developer guide.  This should be a guide targeting a user who simply wants to run the tool.

Patch-set: 5
Reviewer: Gerrit User 163 <163@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1