Update patch set 6
Patch Set 6: Code-Review-1

(29 comments)

Some detailed comments inline.
I have reservations that none of the IPv6 tests are evaluating the ability of the SUT to pass traffic.  This should be made clear in the description of the test suite that it is only evaluating the HTTP content of the command responses.
In General:
- Remove generic pass/fail text.  
- use canonical references that will be aplicable for Danube also in the future.  (no wiki's, versioned references needed)
- Clarify preconditions on test cases.

Patch-set: 6
Label: Code-Review=-1