Update patch set 14
Patch Set 14:

Hi Georg,

My understanding is that a majority of committers (even a supermajority) voted +1 on patchset 13 or 14 - so I +2'd to enable the merge. Your merge was fine, I think. Future changes should be made in incremental changes. The other Dovetail use-case issues now need to be rebased (they will be in conflict, I believe).

Patch-set: 14