Update patch set 14
Patch Set 14:

Chris: I saw Georg, you, me, Hongbo, Wenjing. Chris, you are still listed as a committer in the Info file, along with Trevor, Zenghui, Lincoln, Leo, and Fu Qiao. You are correct that we need one more vote. I invite one or more of the other committers to comment on this change - either it needs to be backed out (if -1) or the merge will be confirmed (if +1).

Patch-set: 14