Update patch set 14
Patch Set 14:

We talked at the summit about how to proceed with this accidental merge. Since the votes in Gerrit seem to be confusing and incomplete, I'd like to propose having an official and traceable vote on IRC at the next Dovetail meeting to either approve or revert this merge.

Patch-set: 14