Update patch set 11
Patch Set 11:

I'm still not sure what, if anything, specific to IPv6 is tested here. It would be nice to have something in the test specs that is specific to IPv6 (say, example commands for IPv6 subnet creation, assignment of IPv6 addresses to instances/ports + passing some IPv6 traffic).

Patch-set: 11