Update patch set 7
Patch Set 7: Code-Review-1

(18 comments)

All in all, the document needs significantly more work in my view. The description of test scope is a bit unstructured (mixing what is in scope and arguments why certain things are out of scope), and I'm missing a more comprehensive analysis what basic capabilities are so that Dovetail would be able to analyze if currently available tests are sufficient or not. Some more dtailed comments inline.

Patch-set: 7
Label: Code-Review=-1