Update patch set 1
Patch Set 1:

While I have no problem merging this particular patch, I would like to clarify the process for adding a test case to the required tests. The process, as I understood it, was:
1. Propose tests as part of optional test suite - approval requires two committers (+2/merge)
2. Add test case documentation to optional tests - approval means that documentation meets minimum required standards, requires majority of committers
3. Propose addition of tests to mandatory tests - approval requires majority of committers

Is this everyone's understanding? This mentions "mandatory vping test cases" and "proposed_tests" - is proposed_tests the list of required tests for passing Dovetail? In which case, this submission should require a majority of committers.

Patch-set: 1