Update patch set 8
Patch Set 8:

> I agree with Trevor's comment after reading Gabriel's reply to him.
 > A Stress test and a Resiliency test are two different types of
 > tests. Resiliency tests evaluate the ability of a system to recover
 > from difficult circumstances, both during the difficulty and
 > *after* the difficulty has been removed (operation should return to
 > normal). At least, we should agree on the (different) goals of
 > stress tests and resiliency tests.

Hi Morton, really appreciate your comment! I see there is a difference between the understanding of stress testing and resiliency testing. I think it is necessary we align with the understanding first. In my view, I wouldn't make stress testing and resiliency testing into totally different classes of testing regarding the test methodologies. Terminology of resiliency could be found in GS NFV 003 which is "ability of the NFV framework to limit disruption and return to normal or at a minimum acceptable service
delivery level in the face of a fault, failure, or an event that disrupts the normal operation". 

Apparently, resiliency testing evaluates the ability of SUT's recoverability in face of fault, failure or disrupts in normal operation or simply the ability of SUT absorbing "disruptions".  This is exactly what stress testing does. Only that resiliency testing has wider scope than stress testing in terms that resiliency testing could take different forms of inputs that causing "disruptions" and stress testing take some kind of "load" as input. These kinds of "load" will cause "disruption" which could be easily found in system logs. It is the purpose to raise such "load" to the SUT. In Danube, we only considered the resiliency test with "load" input.

I guess the "difficulty" you mentioned in your comment is a synonym for "disruption". The difficult circumstances could be caused by different forms of inputs. The system should have the ability to deal with the difficulties during the test and return to normal. It is the goal of the proposed test to evaluate the SUT if it could provide an acceptable level of capacity during difficult circumstances. It is the reason why the test case contains two subtests to do secondary validation that if the system has the ability to handle whatever "disruption" occurs during the test. I think I should make the discription of this class of tests more clear to avoid any misleading information.

In addition, I just found out, couple of days ago through email discussion, that you are the rapporteur for NFV TST009 which is started in May and might investigate stress testing. I planned to discuss with you to gain more insights. Moreover, if you also need some personal experience of mine there, I would glad to offer some help too!

Patch-set: 8