Update patch set 8
Patch Set 8:

Our original proposal re. introducing stress tests included loading the data-path in addition to loading the system with VNF lifecycle events https://etherpad.opnfv.org/p/DanubeStressTest. The work done to date is valuable but I think we are some ways from having a set of representative stress or resiliency tests with a stable baseline of results suitable for a commercial compliance program. In addition the scope of the initial CVP release is limited to functional testing.

Patch-set: 8