Update patch set 8
Patch Set 8: -Code-Review

Hi Gabriel, thanks for your reply. It took a bit longer than I hoped to respond.  I understand what you described in you reply, but I have to say much of what you described does not come through in the file we are reviewing.
For example, you have a reference to the bottleneck Stress test page.
When you say "The system should have the ability to deal with the difficulties during the test and return to normal.", are you assuming that the Bottleneck TC1 for baseline Throughput is running in the background when you create and destroy *N1* VMs?  If so, you could evaluate the three phases of resiliency: before, during, and *after* the stressful circumstances.
So, there are (at least) three results for every test, and the before and after test phases should produce the *same* performance.
FYI - we don't plan to specify resiliency tests in TST009, but the Baseline test would benefit from improved methods to measure Throughput (eventually).

Patch-set: 8