Update patch set 8
Patch Set 8:

> Hi Gabriel, thanks for your reply. It took a bit longer than I
 > hoped to respond.  I understand what you described in you reply,
 > but I have to say much of what you described does not come through
 > in the file we are reviewing.
 > For example, you have a reference to the bottleneck Stress test
 > page.
 > When you say "The system should have the ability to deal with the
 > difficulties during the test and return to normal.", are you
 > assuming that the Bottleneck TC1 for baseline Throughput is running
 > in the background when you create and destroy *N1* VMs?  If so, you
 > could evaluate the three phases of resiliency: before, during, and
 > *after* the stressful circumstances.
 > So, there are (at least) three results for every test, and the
 > before and after test phases should produce the *same* performance.
 > FYI - we don't plan to specify resiliency tests in TST009, but the
 > Baseline test would benefit from improved methods to measure
 > Throughput (eventually).

Hi Morton, thanks for the detailed explanation! Really appreciate it!! Yes, we should make clarification in details for the terminology and methodology of stress/resiliency test. I fully agree with you that if we perform Bottlenecks TC1 running in the background, we should provide the throuphput results before, during and after concurrently life-cycle requests. In the proposed test, we do not perform the throughput baseline test. We just use vPing to verify the connectivity between a VM pair. This test only verify if the system could correctly handles each life-cycle request without involving dataplane evaluation.

As to the throughput baseline test that not included and explained here, I could provide more details to you if you need it (regarding the influence of choices of parameters, stopping criteria and planning tests based on it, etc.)

Patch-set: 8