Update patch set 4
Patch Set 4: Code-Review+1

Re: Dave's comment. The paragraph was actually in the previous version too. Georg moved it out of the enumerated analysis section to make it a bit clearer that it's for future consideration. A lot of people did express desire to see these considerations. I would recommend we keep it. Thanks for your willingness of going with it.

Patch-set: 4
Label: Code-Review=+1