Update patch set 3
Patch Set 3:

> From the jenkins log for the failure:
 > 
 > ERROR: InvocationError: '/home/jenkins/opnfv/slave_root/workspace/dovetail-verify-master/.tox/pep8/bin/flake8
 > /home/jenkins/opnfv/slave_root/workspace/dovetail-verify-master'
 > 
 > Is it possible there's a build error?

Hi Eddie, I guess this is because of the update of pep8 check.
All statements 'except:' in the source code are regarded as error and needs to be 'except Exception:'.
I am trying to ignore this error in our pep8 check temporarily.
And in the next release we can consider to fix these.

Patch-set: 3
Label: Verified=0