Update patch set 1
Patch Set 1:

One of the Jenkins errors is due to a Tox issue Xudan is fixing in DOVETAIL-537. The other error looks to be an error from anteater because of the binary image files I updated. I probably need to get someone to add an exception for this. Any guidance is appreciated

Patch-set: 1
Reviewer: Gerrit User 7047 <7047@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0