Update patch set 2
Patch Set 2:

> > the names now can be used in the dovetail command line option
 > > "--testcase" directly, is that correct? thanks. right thing to do
 > > btw.
 > 
 > Sorry, Dovetail don't support option '--testcase'. The smallest
 > unit of testing now is testarea not testcase. The reason I added
 > the testcase name in the repo into docs is that it can help users
 > to identify the test cases from both the repo and docs sides.

can we add a new feature in F to support executing a specific test case? sometimes and it is quite often that we need to rerun a specific test case to do the troubleshooting, currently it is very difficult, we have to comment out all the other test cases.

Patch-set: 2