Remove approve / no approve options for reviewers
Distinguish results shared or submitted from a tester

JIRA: DOVETAIL-544
JIRA: DOVETAIL-545

[cvp-web]There is consensus within C&C committee to move towards a
simpler means of voting by reviewers. We will use email voting (+1) on
the [cvp@opnfv.org|mailto:cvp@opnfv.org] email alias. For now, can we
disable the approve / not approve options in the drop-down in the
'Operation' column of the results? We can revisit this feature in the
future after some experience with the manual method.

[cvp-web]When a user shares or submits results, the recipient of
those results (share target or reviewer) do not know where the results
originated from. To remedy this issue, we will add a column labeled
'Owner' to the 'My Results' table that shows the email address or the
sending party?

Change-Id: I586b01ad113d898fe3d5824025df9877118fa622
Signed-off-by: grakiss <grakiss.wanglei@huawei.com>