Update patch set 2
Patch Set 2:

(6 comments)

I went through the changes and it seems there are some inconsistencies between the test cases listes in the rst files and the test cases listed in the yaml. Can you please check?

This points to one thing we should improve on: having a singular machine and human readable reference for the test cases consumed by both the client and the web portal.

Patch-set: 2