Update patch set 2
Patch Set 2:

(6 comments)

Hi Georg, Thanks very much for your careful review. I have checked that all the 205 test cases left are all the ones actually executed by Dovetail now. In order to keep the sequence of the test cases, I replaced the alias test cases with the ones that will be executed by Tempest 14.0.0 and remove the redundancy ones. So for some test cases, you may think that I removed them.
I have reply your comment on Leo's patch anout the testcases.json. In my opinion, that testcases.json is suitable for both the client and the web. If it can be accepted by others, then it can also be used by the client in its next release.

Patch-set: 2