[ovp-web] Use new program logo/text for OVP portal home page
JIRA: DOVETAIL-570

OPNFV marketing has provided new logos for OVP that will be incorporated
on the portal home page and attached to this issue. The 'Infrastructure'
logo must by used. The other two logos (Application, Lab) currently on
the development site must be removed for now.

Position the new logo in the upper-left side and position the blurb
language to the right of it. The new blurb language to swap in (all
uniform, non-bold font) is provided below:

    The OPNFV Verified program (OVP) demonstrates the readiness and
    availability of commercial products based on OPNFV. Verified products
    and services submitted by vendors and service providers become compliant
    by implementing explicitly defined interfaces, behaviors and key
    features while retaining distinct and value-added innovations across
    features and capabilities. The OPNFV Verified program grants products
    and services with the "OPNFV Verified" mark. Get started with your OVP
    application today.

Change-Id: Iaf4172fe61ea46133c474207f77ba24b841211e4
Signed-off-by: grakiss <grakiss.wanglei@huawei.com>