[ovp-web] Add SUT version number to OVP portal
JIRA: DOVETAIL-572

We currently have a label column in the OVP test result view. This
allows users to attach a custom label field to their results. It is at
the users discretion to fill out this field.
We need a separate SUT version column to track the product version
number being submitted for official review. The field can be left blank
when users upload and share results in a private manner. However, the
OVP portal must check that this new 'SUT version' field is populated
before results are officially submitted for review. An error message
prompting the user to fill out this field if it is blank must be
implemented when users submit results for review.

Change-Id: I1ca915423a56e8173a3a13f1b84b066cf91d9291
Signed-off-by: chenjiankun <chenjiankun1@huawei.com>
(cherry picked from commit af1debe11cf4793bae049ff3ffd63dc906bbc1c9)