Update patch set 1
Patch Set 1:

I have added Brandon to the list of reviewers for additional input from marketing perspective.

I assume official marketing will use "OPNFV Verified Program", but as part of the documentation and everyday "verbal use", the shorter version is handy. Anyway, I can easily expand OVP to the full term in the documentation if that is desired.

Patch-set: 1