Update patch set 2
Patch Set 2:

Agree that the redundancy is annoying. But a concise listing of test cases and sub test cases does not seem to exist anywhere else in the docs. It's scattered across the test specs. There are other inconsistencies in naming that we can clean up in the future.

Patch-set: 2
Reviewer: Gerrit User 7047 <7047@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0