Test case description for BGPVPN Tempest API tests
Adding test case descriptions for the BGPVPN API tests which are
implemented as Tempest plugin in the networking-bgpvpn repository. These
tests have already been integrated for Euphrates into Functest test and
are hence a candidate for inclusion in the test scope of the OVP
Euphrates release as extension of the existing SDNVPN test cases.

JIRA: DOVETAIL-598

Change-Id: Id8578ed1e198ddbde967793737c69a5940c22475
Signed-off-by: Georg Kunz <georg.kunz@ericsson.com>