Update patch set 5
Patch Set 5:

I agree that it makes sense to keep the report option even if the local db push is not required for OVP portal results. The user may not want to have tarballs created for trial/debug test runs.

The system command I used yields the docker0 bridge IP, which essentially proxies for the host IP. Are you observing something different?

Patch-set: 5