Update patch set 5
Patch Set 5:

Do we need the "default" argument to the report option? Logically, the command line option could look like this:

dovetail ... --report [URL]

Here the URL is optional. If it is empty, build the default URL; if it is given, use that one. I haven't looked at the code yet, though.

Patch-set: 5