Update patch set 1
Patch Set 1:

I noticed new anteater output in a previous commit. It's complaining about the usage of shell=True in a python subprocess command, as a security concern. Are we really expected to waste time modifying these commands in previously checked in code that has nothing to do with this patch? Security should not be an issue in test tool code. Almost every patch in Gerrit I've made has resulted in dumb time wasting errors - swimming in molasses.

Patch-set: 1
Label: Verified=0