Update patch set 2
Patch Set 2:

Hi Dave,

very good point. After completing the patchset in terms of required data, I'd like to open the discussion to the entire team. Everyone, please chime in. This is an important governance issue of the project.

I very much agree that a balance among companies is important for Dovetail. I nevertheless wanted to bring these two promotions forward for discussion in the team based on the recent and current contributions of both Eddie and Xudan (see commit message). As mentioned above, we need to retain a balance among companies in Dovetail and revisiting the entire list of committers is one way of approaching this.

I wanted to propose to split up this patchset into two, so that we can vote individually on each promotion, but Eddie has already stated that he does not pursue a promotion at this point in time. I'll nevertheless leave the patchset as is for further discussion for some time.

Patch-set: 2