Update patch set 2
Patch Set 2:

> Please remove me from consideration. I couldn't care less about the
 > moniker and the superfluous ceremony. But please add Xudan. It's
 > laughable she's not a 'committer' even though she has more commits
 > than anyone else by a long shot. She's accountable without being
 > asked or having any authority.

That wasn't the point, Eddie, and I did not mean any offense. This is not just pomp and ceremony, this project is the manifestation of the compliance test suite for OPNFV, and as such, committer make-up is an important check-and-balance to scope creep. It is also the tool used to run that test suite, and in that case, different checks and balances are appropriate (and of course frequent contributors should be committers in this case).

The first role (check-and-balance to test suite scope) is where organizational diversity is important. I would suggest that you, Leo, Xu Dan, Leo, and Wenjing discuss whether the frequent contributors are also the best people to represent HUawei's interests in the scope discussions.

For my part, as I said while we were adding committers in the past, my intention is to resign as a committer. I think now, once the first release has happened, is the right time for that to happen, as Zenghui is more active on the project than I am.

Patch-set: 2
Reviewer: Gerrit User 514 <514@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0