Update patch set 4
Patch Set 4: -Code-Review

> Hi Xu Dan,
 > interesting. For clarification: your local environment does not
 > support BGPVPN, so the tests are expected to fail?

Yes. It does not support BGPVPN.
So for this test case, it needs to check the value of 'status' rather than 'cirteria', right? If so, Dovetail needs to do some adaption for this test case because for all other test cases dovetail checks 'criteria'.

Patch-set: 4
Label: Code-Review=0