Update patch set 2
Patch Set 2:

Yes, the thinking is to allow the user to execute before running the compliance suites as a pre-flight check.

In a separate patch, I'm thinking to add a subset of these tests into the stop on failure flag/logic. If the flag is set, the connection_test (since the run duration is short) would be executed in between test areas. If it failed and the flag is set, test execution would stop.

I can modify this patch to put it into a separate healthcheck test suite and possibly leave it in the debug test suite also.

Patch-set: 2