Update patch set 3
Patch Set 3: Code-Review+1

(2 comments)

Overall, this looks good.  I am just looking for ways to make this easier and not add additional steps to get dovetail to run.  I removing the strict check for the duration of the waiver for all participants that run the test and using the log files that are produced to clarify whether strict API will pass.  This makes one process for all submissions with less options to document and or tweak.

Patch-set: 3
Reviewer: Gerrit User 6769 <6769@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=+1