Update patch set 2
Patch Set 2:

Hi Cedric. The next release of Dovetail/OVP will likely be based on Fraser (not finally decided yet). Then the gap will be smaller.

As you mentioned, this exemption is a special case. So, even with a smaller gap, there will not be an official mechanism supported by upstream OpenStack to help us (and rightfully so from an OpenStack perspective). Given that this exemption is designed to be a temporary thing, maintaining a patch in Dovetail seems the easiest option to me - which is also not impacting Functest but the effort is solely on us.

Patch-set: 2