Renaming test cases for better readability
This patchset proposes a new naming scheme for test cases. The ides is
to replace the current numbered scheme

  <capability>.tc<number>

with:

  <category>.<capability/test case description>

The motivation for this is that a the numbering does not convey any
information about what the test case is doing and hides to which test
case implementation in Functest or Yardstick a test case spec refers.

Note: this patchset has not yet been tested in Dovetail itself. Its goal
is start a discussion about the naming scheme.

JIRA: DOVETAIL-647

Change-Id: If33b4bf88e451b372c2af10895ff92a87d9d961d
Signed-off-by: Georg Kunz <georg.kunz@ericsson.com>