Update patch set 1
Patch Set 1:

the idea behind restructuring the IPv6 test cases is to handle them just like the other tempest tests. All IPv6 test cases are consolidated in a dedicated category, just like compute and volume. Of course, you can argue that IPv6 is a subcategory of tempest.networking, but I personally think it is fine for us to keep them separate.

One other aspect that just comes to my mind: we need to make sure that we can distinguish mandatory and optional test cases and handle them differently.

Patch-set: 1