Extend web portal to show if API exemption was used during a test run
JIRA: DOVETAIL-634

The OVP web port needs to read from the test results if the strict API
response validation exemption was used or not and show it
correspondingly.

Detailed description:

    As discussed on the Dovetail call on 2018-04-11, the web portal should
show if API response validation was enabled or disabled. The information
if API response validation was enabled or disabled is to be determined
based on the dovetail.log file. If API response validation was
disabled,** the log file contains the line "Strict API response
validation DISABLED.". Upon uploading a result tar-package, the web
portal should extract this information from the log file, store it in
the database, and display it as follows:

    This indication shall be displayed along with the test summary on the
Test Run Results page as either:

    "API response validation enabled"   or
    "API response validation disabled"

    For clarification: API response validation is enabled by default. API
response validation is disabled only if  the --no-api-validation command
line option is given.

Change-Id: Ifacb1eb2127cc8dc6364c98dead1217aeedfd359
Signed-off-by: chenjiankun <chenjiankun1@huawei.com>
(cherry picked from commit 67de42416576c952ec0d6a159a0d57fc4cabc414)